+++ telegram-adium/TelegramAccountViewController.m
@@ -1,17 +1,19 @@
-/* 
- * Adium is the legal property of its developers, whose names are listed in the copyright file included
- * with this source distribution.
- * 
- * This program is free software; you can redistribute it and/or modify it under the terms of the GNU
- * General Public License as published by the Free Software Foundation; either version 2 of the License,
- * or (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General
- * Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License along with this program; if not,
- * write to the Free Software Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+/**
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02111-1301  USA
+ *
+ * Copyright Matthias Jentsch 2014-2015
  */
 
 #import "TelegramAccountViewController.h"
@@ -20,7 +22,7 @@
 #import <Adium/AIService.h>
 #import <AIUtilities/AIStringFormatter.h>
 #import <AIUtilities/AIAttributedStringAdditions.h>
-#import <AIUtilities/AIPopUpButtonAdditions.h>
+#import <AIUtilities/AIPopUpButtonAdditions.h>DISPLAY
 
 #include "telegram-purple.h"
 
@@ -46,11 +48,18 @@
   }
   [radio_Encryption selectCellAtRow:row column:0];
   
+  NSString *passwordTwoFactor = [account preferenceForKey:@"Telegram:"TGP_KEY_PASSWORD_TWO_FACTOR
+                                 group:GROUP_ACCOUNT_STATUS] ?: nil;
+  [textField_password setStringValue:passwordTwoFactor];
+  
   id s = [account preferenceForKey:@"Telegram:"TGP_KEY_HISTORY_SYNC_ALL group:GROUP_ACCOUNT_STATUS];
   [checkbox_historySyncAll setState:[s boolValue]];
   
-  id s4 = [account preferenceForKey:@"Telegram:"TGP_KEY_DISPLAY_READ_NOTIFICATIONS group:GROUP_ACCOUNT_STATUS];
-  [checkbox_displayReadNotifications setState:[s4 boolValue]];
+  id read = [account preferenceForKey:@"Telegram:"TGP_KEY_DISPLAY_READ_NOTIFICATIONS group:GROUP_ACCOUNT_STATUS];
+  [checkbox_displayReadNotifications setState:[read boolValue]];
+  
+  id send = [account preferenceForKey:@"Telegram:"TGP_KEY_SEND_READ_NOTIFICATIONS group:GROUP_ACCOUNT_STATUS];
+  [checkbox_sendReadNotifications setState:[send boolValue]];
 
   NSString *inactiveDaysOffline = [account
                                    preferenceForKey:@"Telegram:"TGP_KEY_INACTIVE_DAYS_OFFLINE
@@ -69,6 +78,10 @@
 	
   NSArray *selections = @[@"ask", @"always", @"never"];
   
+  [account setPreference:[textField_password stringValue]
+                  forKey:@"Telegram:"TGP_KEY_PASSWORD_TWO_FACTOR
+                   group:GROUP_ACCOUNT_STATUS];
+  
   [account setPreference:selections[[radio_Encryption selectedRow]]
                   forKey:@"Telegram:"TGP_KEY_ACCEPT_SECRET_CHATS
                    group:GROUP_ACCOUNT_STATUS];
@@ -79,6 +92,10 @@
   
   [account setPreference:[NSNumber numberWithBool:[checkbox_displayReadNotifications state]]
                   forKey:@"Telegram:"TGP_KEY_DISPLAY_READ_NOTIFICATIONS
+                   group:GROUP_ACCOUNT_STATUS];
+  
+  [account setPreference:[NSNumber numberWithBool: [checkbox_sendReadNotifications state]]
+                  forKey:@"Telegram:"TGP_KEY_SEND_READ_NOTIFICATIONS
                    group:GROUP_ACCOUNT_STATUS];
   
   [account setPreference:[textField_historyRetrieveDays stringValue]