+++ telegram-base.h
@@ -22,6 +22,12 @@
 
 #include "telegram-purple.h"
 
+struct request_password_data {
+  struct tgl_state *TLS;
+  void (*callback)(struct tgl_state *TLS, const char *string, void *arg);
+  void *arg;
+};
+
 void read_state_file (struct tgl_state *TLS);
 void read_auth_file (struct tgl_state *TLS);
 void write_auth_file (struct tgl_state *TLS);
@@ -33,7 +39,14 @@ void write_secret_chat_gw (struct tgl_state *TLS, void *extra, int success, stru
 
 void telegram_login (struct tgl_state *TLS);
 void request_code_entered (gpointer data, const gchar *code);
-int generate_ident_icon(struct tgl_state *TLS, unsigned char* sha1_key);
+void request_password (struct tgl_state *TLS, void (*callback)(struct tgl_state *TLS, const char *string, void *arg), void *arg);
 
 void request_accept_secret_chat (struct tgl_state *TLS, struct tgl_secret_chat *U);
+
+gchar *get_config_dir (struct tgl_state *TLS, char const *username);
+gchar *get_download_dir (struct tgl_state *TLS);
+void assert_file_exists (PurpleConnection *gc, const char *filepath, const char *format);
+
+int tgp_visualize_key(struct tgl_state *TLS, unsigned char* sha1_key);
+
 #endif