+++ tgp-utils.c
@@ -18,8 +18,18 @@
  Copyright Matthias Jentsch 2014-2015
  */
 
+#ifdef HAVE_CONFIG_H
+#include "config.h"
+#endif
+
 #include "tgp-utils.h"
+#include "msglog.h"
+#include "lodepng/lodepng.h"
+
 #include <purple.h>
+#ifdef HAVE_LIBWEBP
+#include <webp/decode.h>
+#endif
 
 connection_data *get_conn_from_buddy (PurpleBuddy *buddy) {
   connection_data *c = purple_connection_get_protocol_data (
@@ -68,12 +78,12 @@ int str_not_empty (const char *string) {
   return string && string[0] != '\0';
 }
 
-int our_msg (struct tgl_state *TLS, struct tgl_message *M) {
-  return (M->flags & FLAG_SESSION_OUTBOUND) != 0;
+int tgp_outgoing_msg (struct tgl_state *TLS, struct tgl_message *M) {
+  return (M->flags & TGLMF_SESSION_OUTBOUND);
 }
 
-int out_msg (struct tgl_state *TLS, struct tgl_message *M) {
-  return M->out;
+int tgp_our_msg (struct tgl_state *TLS, struct tgl_message *M) {
+  return TLS->our_id == tgl_get_peer_id(M->from_id);
 }
 
 tgl_peer_t *find_peer_by_name (struct tgl_state *TLS, const char *who) {
@@ -124,3 +134,38 @@ void tgp_g_list_free_full (GList *list, GDestroyNotify free_func) {
   g_list_foreach (list, (GFunc)free_func, NULL);
   g_list_free (list);
 }
+
+const char *tgp_mime_to_filetype (const char *mime) {
+  int len = (int) strlen (mime);
+  int i;
+  for (i = 0; i < len - 1; i ++) {
+    if (mime[i] == '/') {
+      return mime + i + 1;
+    }
+  }
+  return NULL;
+}
+
+#ifdef HAVE_LIBWEBP
+void *tgp_webp_load_png (const char *filename, size_t *size) {
+  gchar *data = NULL;
+  size_t len;
+  GError *err = NULL;
+  g_file_get_contents (filename, &data, &len, &err);
+  if (err) { warning ("cannot open file %s: %s.", filename, err->message); return NULL; }
+  
+  int width, height;
+  uint8_t* decoded = WebPDecodeRGBA ((const uint8_t*)data, len, &width, &height);
+  g_free (data);
+  if (! decoded) { warning ("failed decoding webp"); return NULL; }
+  
+  unsigned char* png = NULL;
+  size_t pnglen;
+  unsigned error = lodepng_encode32 (&png, &pnglen, decoded, width, height);
+  free (decoded);
+  if (error) { warning ("failed encoding png"); return NULL; }
+  
+  *size = pnglen;
+  return png;
+}
+#endif