+++ tgp-utils.h
@@ -34,12 +34,12 @@ tgl_peer_t *find_peer_by_name (struct tgl_state *TLS, const char *who);
 /**
  * Return whether this message was created by our client in this session
  */
-int our_msg (struct tgl_state *TLS, struct tgl_message *M);
+int tgp_outgoing_msg (struct tgl_state *TLS, struct tgl_message *M);
 
 /**
- * Return whether this message was created by the current user.
+ * Return whether this message was created by the current user
  */
-int out_msg (struct tgl_state *TLS, struct tgl_message *M);
+int tgp_our_msg (struct tgl_state *TLS, struct tgl_message *M);
 
 const char *format_time (time_t date);
 char *format_img_full (int imgstore);
@@ -49,5 +49,7 @@ long tgp_time_n_days_ago (int days);
 char *tgp_g_format_size (gint64 size);
 void tgp_g_queue_free_full (GQueue *queue, GDestroyNotify free_func);
 void tgp_g_list_free_full (GList *list, GDestroyNotify free_func);
+const char *tgp_mime_to_filetype (const char *mime);
+void *tgp_webp_load_png (const char *filename, size_t *size);
 
 #endif